+++ Unity/Assets/Model/Module/Numeric/INumericWatcher.cs
@@ -2,6 +2,6 @@
 {
 	public interface INumericWatcher
 	{
-		void Run(long id, int value);
+		void Run(long id, long value);
 	}
 }
